requirements.txt
# Bot
botbuilder-integration-aiohttp>=4.15.0

# Server
quart==0.20.0
uvicorn==0.38.0
gunicorn==23.0.0

# Agent
llama-index==0.14.7
llama-index-llms-bedrock-converse==0.11.0

# Google
google-ads==28.4.0

# Azure Storage
azure-data-tables==12.7.0

# Helpers
openpyxl==3.1.5
python-docx==1.2.0
pdfplumber==0.11.8